Persistent session via cookie pt 1
Works with the general concept. Provides the ability to skip re-logging in
when a cookie exists in the session or is provided with the request. At
this point it probably only lives as long as the session.
TODO: Persist cookie data to database. Make cookie live longer (a week?).
Create a cookie handler class so various cookie related methods don't live
in the LoginServlet or the LoginFilter. Finish the CookieService class and
create the CookieDao class.